refactor(eval): split evaluation metrics modules
